+++ lithos-lib/src/main/scala/stratum/message/Announcement.java
@@ -1,14 +1,17 @@
 package stratum.message;
 
+import com.redbottledesign.bitcoin.rpc.stratum.message.Result;
+import org.json.JSONArray;
 import stratum.BlockTemplate;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 
 import static stratum.Utils.jsonArray;
 
 public class Announcement {
 
-	public static AnnouncementMessage difficulty(BigInteger difficulty) {
+	public static AnnouncementMessage difficulty(BigDecimal difficulty) {
 		return new AnnouncementMessage("mining.set_difficulty", () -> jsonArray(difficulty));
 	}
